requirements.txt
elasticsearch
alembic==1.4.3
bcrypt==3.2.0
blinker==1.4
captcha==0.3
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coinbase-commerce==1.0.1
cryptography==3.2.1
cssmin==0.2.0
dnspython==2.0.0
dominate==2.6.0
elasticsearch==7.10.0
email-validator==1.1.1
Flask==1.1.2
Flask-Assets==2.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.9.0
Flask-DebugToolbar==0.11.0
Flask-Limit==1.0.2
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
flask-msearch==0.2.9
flask-session-captcha==1.2.0
Flask-Sessionstore==0.4.5
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
flask-whooshalchemy3==0.2.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
limits==1.5.1
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
passlib==1.7.4
pbr==3.1.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyMySQL==0.10.1
python-dateutil==2.8.1
python-editor==1.0.4
redis==3.5.3
requests==2.25.0
selenium==3.141.0
sentry-sdk==0.19.4
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-FullText-Search==0.2.5
urllib3==1.26.2
uWSGI==2.0.19.1
visitor==0.1.3
webassets==2.0
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.3
zipp==3.4.0
